# cql3: Remove term, replace with expression