--- a/cdc/log.cc
+++ b/cdc/log.cc
@@ -44,7 +44,6 @@
 #include "types/tuple.hh"
 #include "cql3/statements/select_statement.hh"
 #include "cql3/multi_column_relation.hh"
-#include "cql3/tuples.hh"
 #include "cql3/untyped_result_set.hh"
 #include "log.hh"
 #include "utils/rjson.hh"