--- a/cql3/constants.cc
+++ b/cql3/constants.cc
@@ -62,7 +62,7 @@ expr::expression constants::marker::to_expression() {
     return expr::bind_variable {
         .shape = expr::bind_variable::shape_type::scalar,
         .bind_index = _bind_index,
-        .value_type = _receiver->type
+        .receiver = _receiver
     };
 }