--- a/cql3/statements/update_statement.hh
+++ b/cql3/statements/update_statement.hh
@@ -48,8 +48,6 @@
 
 namespace cql3 {
 
-class term;
-
 namespace statements {
 
 /**