@cvybhu 2021-10-15
v2:
* Moved `expr::visit` to a separate commit
* Fixed recursing in `fill_prepare_context`
* Bring back casting `optional` to `bool` in attributes
* Remove timeout check
* Fix `term.hh` include
* Added `find_expression` and used it to implement `find_atom` and `contains_bind_marker`

All commits pass `dev-test`, the branch passes the `cql_additional_tests.py:TestCQL` dtest