@avikivity 2021-10-17
Looks better now (mostly due to my improved understanding of the series). Please also measure performance and add that to the cover letter.