@avikivity 2021-10-31
> 
> * On master ([a21b1fb](https://github.com/scylladb/scylla/commit/a21b1fbb2f481671c80560d96a8acc1cf916b468)) I get:
>   ```
>   176506.60 tps ( 77.0 allocs/op,  12.0 tasks/op,   45831 insns/op)
>   
>   median 176506.60 tps ( 77.0 allocs/op,  12.0 tasks/op,   45831 insns/op)
>   median absolute deviation: 0.00
>   maximum: 176506.60
>   minimum: 176506.60
>   ```
> * On this branch I get:
>   ```
>   172225.30 tps ( 75.1 allocs/op,  12.1 tasks/op,   46106 insns/op)
>   
>   median 172225.30 tps ( 75.1 allocs/op,  12.1 tasks/op,   46106 insns/op)
>   median absolute deviation: 0.00
>   maximum: 172225.30
>   minimum: 172225.30
>   ```

Any idea where the regression comes from? There are fewer allocations, but more instructions.