@avikivity 2021-10-31
Looks good, but I think you missed recursion into binary_operator::rhs.