@cvybhu 2021-11-05
> Any idea where the regression comes from? There are fewer allocations, but more instructions.

I'm not sure, some guesses would be:
* More copies occurring - things that used to be cloning a shared pointer are now copying data - this would cause more instructions to be executed.
* There could be places where `std::move` is missing - copying a `shared_ptr` was cheap, but now it's an `expression`. I tried to add `std::move` everywhere, but there still might be some places where it's missing.
* Maybe it's just different code doing a bit different things, so the performance is different

I tried making some flamegraphs (with 100x operations-per-shard and `-fno-omit-frame-pointer`), but didn't notice anything:
https://github.com/cvybhu/attached-files/tree/main/term2-flame

Is there some other method I could use to pinpoint where the performance is going?