@avikivity 2021-11-07
> > Any idea where the regression comes from? There are fewer allocations, but more instructions.
> 
> I'm not sure, some guesses would be:
> 
> * More copies occurring - things that used to be cloning a shared pointer are now copying data - this would cause more instructions to be executed.

But why? This should happen only during prepare.

> * There could be places where `std::move` is missing - copying a `shared_ptr` was cheap, but now it's an `expression`. I tried to add `std::move` everywhere, but there still might be some places where it's missing.
> * Maybe it's just different code doing a bit different things, so the performance is different
> 
> I tried making some flamegraphs (with 100x operations-per-shard and `-fno-omit-frame-pointer`), but didn't notice anything: https://github.com/cvybhu/attached-files/tree/main/term2-flame
> 
> Is there some other method I could use to pinpoint where the performance is going?

Run the benchmark on individual commits and see where it goes wrong, that will narrow the scope.

Another option is `perf diff` (perhaps on the last good and first bad commits, to minimize code differences.