@cvybhu 2021-11-12
I made some benchmarks:
https://github.com/cvybhu/attached-files/tree/main/term2-pr3-bench

But strange things seems to be happening. 
 * The commit that breaks performance the most is this one: c84e941df99b64f17e27a2bbf8172d805023c352
    But it doesn't make any sense - there are no sets in `perf_simple_query`, it shouldn't be executed at all.
 * Then there is commit dcd3199037457158f6e2516231906d04f9637da6, which only renames `prepare_term` to `prepare_expression`, but this results in a significant performance improvement 0_o.

Here are a few ideas what could maybe cause flaky benchmarks:

* I ran all benchmarks on my work laptop with Fedora 34 and KDE, the only program runnning were the benchmarks.
  Despite this there is some variation in results. I ran the same executable binary 6 times and got these results:
  ```
  180245.85 tps ( 75.0 allocs/op,  12.1 tasks/op,   46035 insns/op)
  182221.58 tps ( 75.0 allocs/op,  12.1 tasks/op,   46020 insns/op)
  181587.81 tps ( 75.0 allocs/op,  12.1 tasks/op,   46117 insns/op)
  178415.41 tps ( 75.0 allocs/op,  12.1 tasks/op,   46033 insns/op)
  181755.36 tps ( 75.0 allocs/op,  12.1 tasks/op,   46107 insns/op)
  180017.17 tps ( 75.0 allocs/op,  12.1 tasks/op,   46017 insns/op)
  ```
  I should probably average a few runs for more consistent results. Although the benchmarks were ran 3 times (for 3 workloads) and the results were pretty much the same on each run.
  
  Maybe I could also try to find another machine for benchmarking and install a headless fedora there.

* Maybe there is something strange happening with builds and caching. There could be a scenario like this:
    * commit `#0` is built - there are some build files created and put in the cache
    * commit `#1` is built - build system reuses these build files
    * commit `#2` is built - it does some big changes (like renaming a function) and invalidates cache, build system rebuilds everything and optimizes better - commit `#2` caused a speedup!
  
  Doing a clean build for each commit should prevent this.
* Big changes in code cause the compiler to slightly change it's optimization decisions and choose other tradeoffs

I will try to make new benchmarks where:
* Results are averaged over 5 tries
* Every build is a clean build